move authentication failure to file
