+++ wire-ios-sync-engine/Tests/Source/SessionManager/SessionManagerTests.swift
@@ -457,77 +457,6 @@ extension IntegrationTest {
 
 // MARK: - Account Deletion
 
-class SessionManagerTests_AuthenticationFailure: IntegrationTest {
-
-    override func setUp() {
-        super.setUp()
-        createSelfUserAndConversation()
-    }
-
-    func testThatItDeletesTheCookie_OnAuthentictionFailure() {
-        // given
-        XCTAssert(login())
-        XCTAssertTrue(sessionManager!.isSelectedAccountAuthenticated)
-
-        // when
-        let account = sessionManager!.accountManager.selectedAccount!
-        sessionManager?.authenticationInvalidated(NSError(code: .accessTokenExpired, userInfo: nil), accountId: account.userIdentifier)
-
-        // then
-        XCTAssertFalse(sessionManager!.isSelectedAccountAuthenticated)
-    }
-
-    func testThatItDeletesAccount_IfRemoteClientIsDeleted() {
-        // given
-        XCTAssertTrue(login())
-        let account = sessionManager!.accountManager.selectedAccount!
-
-        // when
-        sessionManager?.authenticationInvalidated(NSError(code: .clientDeletedRemotely, userInfo: nil), accountId: account.userIdentifier)
-
-        // then
-        guard let sharedContainer = Bundle.main.appGroupIdentifier.map(FileManager.sharedContainerDirectory) else { return XCTFail() }
-        let accountFolder = CoreDataStack.accountDataFolder(accountIdentifier: account.userIdentifier, applicationContainer: sharedContainer)
-
-        XCTAssertFalse(FileManager.default.fileExists(atPath: accountFolder.path))
-    }
-
-    func testThatItTearsDownActiveUserSession_OnAuthentictionFailure() {
-        // given
-        XCTAssert(login())
-        XCTAssertNotNil(sessionManager?.activeUserSession)
-
-        // when
-        let account = sessionManager!.accountManager.selectedAccount!
-        sessionManager?.authenticationInvalidated(NSError(code: .accessTokenExpired, userInfo: nil), accountId: account.userIdentifier)
-
-        // then
-        XCTAssertNil(sessionManager?.activeUserSession)
-    }
-
-    func testThatItTearsDownBackgroundUserSession_OnAuthentictionFailure() {
-        // given
-        let additionalAccount = Account(userName: "Additional Account", userIdentifier: UUID())
-        sessionManager!.environment.cookieStorage(for: additionalAccount).authenticationCookieData = NSData.secureRandomData(ofLength: 16)
-        sessionManager!.accountManager.addOrUpdate(additionalAccount)
-
-        XCTAssert(login())
-        XCTAssertNotNil(sessionManager?.activeUserSession)
-
-        // load additional account as a background session
-        sessionManager!.withSession(for: additionalAccount, perform: { _ in })
-        XCTAssertTrue(self.waitForAllGroupsToBeEmpty(withTimeout: 0.5))
-        XCTAssertNotNil(sessionManager?.backgroundUserSessions[additionalAccount.userIdentifier])
-
-        // when
-        sessionManager?.authenticationInvalidated(NSError(code: .accessTokenExpired, userInfo: nil), accountId: additionalAccount.userIdentifier)
-
-        // then
-        XCTAssertNil(sessionManager?.backgroundUserSessions[additionalAccount.userIdentifier])
-    }
-
-}
-
 class SessionManagerTests_EncryptionAtRestMigration: IntegrationTest {
 
     override var useInMemoryStore: Bool {
